+++ src/playground_base/playground_base/base/topo/param/FBDiscreteLog2Param.cpp
@@ -14,26 +14,26 @@ FBDiscreteLog2ParamNonRealTime::ValueCount() const
   return valueCount;
 }
 
-float 
-FBDiscreteLog2ParamNonRealTime::PlainToNormalized(int plain) const
+double 
+FBDiscreteLog2ParamNonRealTime::PlainToNormalized(double plain) const
 {
   plain = std::bit_width((unsigned int)plain) - 1;
-  return std::clamp(plain / (valueCount - 1.0f), 0.0f, 1.0f);
+  return std::clamp(plain / (valueCount - 1.0), 0.0, 1.0);
 }
 
-int
-FBDiscreteLog2ParamNonRealTime::NormalizedToPlain(float normalized) const
+double
+FBDiscreteLog2ParamNonRealTime::NormalizedToPlain(double normalized) const
 {
-  return FBDiscreteLog2ParamRealTime::NormalizedToPlain(normalized);
+  return FBDiscreteLog2ParamRealTime::NormalizedToPlain((float)normalized);
 }
 
 std::string
-FBDiscreteLog2ParamNonRealTime::PlainToText(FBValueTextDisplay display, int plain) const
+FBDiscreteLog2ParamNonRealTime::PlainToText(FBValueTextDisplay display, double plain) const
 {
   return std::to_string(plain);
 }
 
-std::optional<int>
+std::optional<double>
 FBDiscreteLog2ParamNonRealTime::TextToPlain(FBValueTextDisplay display, std::string const& text) const
 {
   char* end;
