+++ src/playground_base/playground_base/base/topo/param/FBLog2Param.hpp
@@ -33,10 +33,10 @@ public FBLog2ParamRealTime,
 public FBContinuousParamNonRealTime
 {
   FB_NOCOPY_NOMOVE_DEFCTOR(FBLog2ParamNonRealTime);
-  float PlainToNormalized(float plain) const override;
-  float NormalizedToPlain(float normalized) const override;
-  std::string PlainToText(FBValueTextDisplay display, float plain) const override;
-  std::optional<float> TextToPlain(FBValueTextDisplay display, std::string const& text) const override;
+  double PlainToNormalized(double plain) const override;
+  double NormalizedToPlain(double normalized) const override;
+  std::string PlainToText(FBValueTextDisplay display, double plain) const override;
+  std::optional<double> TextToPlain(FBValueTextDisplay display, std::string const& text) const override;
 };
 
 inline float
